backend/pyproject.toml
[tool.poetry]
name = "pindurai-backend"
version = "0.1.0"
description = ""
authors = ["Guionardo Furlan <guionardo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0"
django-ninja = "^1.1.0"
packaging = "^23.2"
pyjwt = "^2.8.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.7"
py2puml = "^0.8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.djlint]
format_css=true
format_js=true
